backend/requirements.txt
Django==4.1.6
psycopg2==2.9.5
Markdown==3.4.1
djangorestframework==3.14.0
django-filter==22.1
drf-spectacular==0.25.1
drf-spectacular-sidecar==2022.12.1
dj-database-url==1.2.0
uWSGI==2.0.21
